+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- `Parser::peek_n()` always returned `None`
+- `Parser::next_{token}` did not work correctly because of `peek_n`
+
 ## [0.5.1] - 2023-03-02
 ### Fixed
 - `Parser::next_string()` did not consume token
@@ -18,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **Breaking Change** `TokenTreeExt::<token>()` now return references
 
-[unreleased]: https://github.com/ModProg/proc-macro-utils/compare/v0.5.0...HEAD
+[unreleased]: https://github.com/ModProg/proc-macro-utils/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/ModProg/proc-macro-utils/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/ModProg/proc-macro-utils/compare/v0.4.0...v0.5.0